fixup! hide behind feature flag
