+++ packages/pass-style/package.json
@@ -34,6 +34,7 @@
     "test": "ava"
   },
   "dependencies": {
+    "@endo/env-options": "^1.1.1",
     "@endo/errors": "^1.1.0",
     "@endo/eventual-send": "^1.1.2",
     "@endo/promise-kit": "^1.0.4",
