+++ packages/pass-style/test/prepare-only-well-formed-strings-passable.js
@@ -1,0 +1,3 @@
+/* global process */
+
+process.env.ONLY_WELL_FORMED_STRINGS_PASSABLE = 'enabled';
